[level 0] Title: 문자열 정렬하기 (1), Time: 0.62 ms, Memory: 75.1 MB -BaekjoonHub

+++ 프로그래머스/lv0/120850. 문자열 정렬하기 （1）/문자열 정렬하기 （1）.java
@@ -4,7 +4,7 @@ class Solution {
         String str = my_string.replaceAll("[^0-9]","");
         int[] answer = new int[str.length()];
         for(int i=0; i<=str.length()-1; i++) {
-            answer[i] = (int)str.charAt(i) - 48;
+            answer[i] = str.charAt(i) - 48;
             //48이 0의 아스키 코드인거임
         }
